dmp/requirements.txt
kafka-python==2.0.2
requests==2.21.0
psycopg2==2.8.2
